Tic-Tac-Toe Game
Here,I use logics for array for both both player X and 0 and by comparing these values to print board where these values are updated at the user-input indexes. Then using the winnning condition, such as [0,1,2] and if the array index level is updated and using these conditions we will return the winner.